INTERLOK-3188 if the version is a snapshot, don't cache "changing modules"
Motivation:

If the version is a snapshot, then we accept that they might change
at any time (for instance you might publish a custom branch to nexus).
In that instance we should force gradle to recheck the dependencies.

Modification:

Add an if statement in the parent based on interlokVersion/interlokUiVersion
which changes the resolution strategy to 0 seconds if they end with
'SNAPSHOT'

Result:

No function change for most users
Users that specify 3.10-SNAPSHOT always check nexus.